pyproject.toml
[project]
name = "3-insightstream"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy<2.0",
    "faiss-cpu==1.7.4",
    "groq>=0.9.0",
    "langchain>=0.1.0",
    "langchain-community",
    "langchain-groq",
    "langchain-huggingface",
    "python-dotenv>=1.0.0",
    "python-magic>=0.4.27",
    "python-magic-bin>=0.4.14",
    "streamlit>=1.30.0",
    "unstructured>=0.12.0",
    "sentence-transformers"
]
